moved functions out of init_shell.c and solved ALL betty errors
