+++ getchar.c
@@ -2,23 +2,27 @@
 #ifndef BUFF_SIZE
 #define BUFF_SIZE 1024
 #endif
+/**
+ * _getchar - reads stream from stdin and stores it to a buffer.
+ * Return: gives the ascii code of character read from input.
+ */
 int _getchar(void)
 {
-    static char buff[BUFF_SIZE];
-    static char *chr;
-    static int pos = 0;  /* New static variable to track position */
-    static int ret = 0;  /* Changed this to static */
+	static char buff[BUFF_SIZE];
+	static char *chr;
+	static int pos;
+	static int ret;
 
-    if (pos >= ret) { /* if all data in buffer has been returned */
-            if ((ret = read(STDIN_FILENO, buff, BUFF_SIZE)) > 0)
-            {
-                    chr = buff;
-                    pos = 0;
-                    return *(chr + pos++); /* return one char and update pos */
-            } else {  /* if no more to read from stdin */
-                    return EOF;
-            }
-    } else { /* if data still in buffer */
-            return *(chr + pos++); /* return one char and update pos */
-    }
+	if (pos >= ret)
+	{
+		ret = read(STDIN_FILENO, buff, BUFF_SIZE);
+		if (ret > 0)
+		{
+			chr = buff;
+			pos = 0;
+			return (*(chr + pos++));
+		} else
+			return (EOF);
+	} else
+		return (*(chr + pos++));
 }
